Extract ebr_impl as separate circ-ebr crate
Move the epoch-based reclamation implementation to a standalone crate,
allowing external projects to use EBR without depending on the full
circ library.
