+++ ebr/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "circ-ebr"
+version = "0.2.0"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+description = "Epoch-based memory reclamation (EBR) for lock-free data structures"
+repository = "https://github.com/kaist-cp/circ"
+
+[dependencies]
+crossbeam-utils = "0.8"
+scopeguard = "1.1.0"
+atomic = "0.5"
+memoffset = "0.7"
+
+[dev-dependencies]
+rand = "0.8"
