+++ src/lib.rs
@@ -1,10 +1,9 @@
 #![doc = include_str!("../README.md")]
 
-pub(crate) mod ebr_impl;
 mod strong;
 mod utils;
 mod weak;
 
-pub use ebr_impl::{cs, Guard};
+pub use circ_ebr::{cs, Guard};
 pub use strong::*;
 pub use weak::*;
